collect_tweets/requirements.txt
tweepy>=4.12.1
toml>=0.10.2
mysql-connector-python